+++ js/constants.js
@@ -1,0 +1,25 @@
+const DESCRIPTIONS = [
+  'Картинка для заставки на телефон',
+  'Красивая фотография',
+  'Милая фотография из галереи',
+  'Популярная картинка из интернета',
+  'Часто скачиваемая картинка',
+];
+
+const NAMES = [
+  'Иван',
+  'Хуан',
+  'Мария',
+  'Кристоф',
+  'Виктор',
+  'Юлия',
+  'Люпита',
+  'Вашингтон',
+];
+
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё',
+];
+
+export {DESCRIPTIONS , NAMES , COMMENTS};
